Apparently you can't have multiple tests in a suite that both override profiles_config_update, or something
