+++ tests/functional/plugins/test_plugins.py
@@ -7,7 +7,21 @@ from dbt.tests.util import (
     run_dbt,
 )
 
-sources_schema_yml = """
+excel_schema_yml = """
+version: 2
+sources:
+  - name: excel_source
+    schema: main
+    meta:
+      plugin: excel
+    tables:
+      - name: excel_file
+        description: "An excel file"
+        meta:
+          external_location: "{test_data_path}/excel_file.xlsx"
+"""
+
+sqlalchemy_schema_yml = """
 version: 2
 sources:
   - name: sql_source
@@ -26,16 +40,20 @@ sources:
           table: "test_table2"
 """
 
-models_source_model1_sql = """
+
+excel1_sql = """
+    select * from {{ source('excel_source', 'excel_file') }}
+"""
+sqlalchemy1_sql = """
     select * from {{ source('sql_source', 'tt1') }}
 """
-models_source_model2_sql = """
+sqlalchemy2_sql = """
     select * from {{ source('sql_source', 'tt2') }}
 """
 
 
 @pytest.mark.skip_profile("buenavista")
-class TestSQLAlchemyPlugin:
+class TestPlugins:
     @pytest.fixture(scope="class")
     def sqlite_test_db(self):
         path = "/tmp/satest.db"
@@ -54,18 +72,22 @@ class TestSQLAlchemyPlugin:
 
     @pytest.fixture(scope="class")
     def profiles_config_update(self, dbt_profile_target, sqlite_test_db):
-        config = {"connection_url": f"sqlite:///{sqlite_test_db}"}
         if "path" not in dbt_profile_target:
             return {}
+
+        config = {"connection_url": f"sqlite:///{sqlite_test_db}"}
+        plugins = [
+            {"name": "excel", "impl": "excel"},
+            {"name": "sql", "impl": "sqlalchemy", "config": config},
+        ]
+
         return {
             "test": {
                 "outputs": {
                     "dev": {
                         "type": "duckdb",
                         "path": dbt_profile_target["path"],
-                        "plugins": [
-                            {"name": "sql", "impl": "sqlalchemy", "config": config}
-                        ],
+                        "plugins": plugins,
                     }
                 },
                 "target": "dev",
@@ -73,29 +95,42 @@ class TestSQLAlchemyPlugin:
         }
 
     @pytest.fixture(scope="class")
-    def models(self):
+    def models(self, test_data_path):
         return {
-            "schema.yml": sources_schema_yml,
-            "source_model1.sql": models_source_model1_sql,
-            "source_model2.sql": models_source_model2_sql,
+            "schema_excel.yml": excel_schema_yml.format(test_data_path=test_data_path),
+            "schema_sqlalchemy.yml": sqlalchemy_schema_yml,
+            "excel.sql": excel1_sql,
+            "sqlalchemy1.sql": sqlalchemy1_sql,
+            "sqlalchemy2.sql": sqlalchemy2_sql,
         }
 
-    def test_sqlalchemy_plugin(self, project):
+    def test_plugins(self, project):
         results = run_dbt()
-        assert len(results) == 2
+        assert len(results) == 3
 
+        # relations_equal
+        check_relations_equal(
+            project.adapter,
+            [
+                "excel_file",
+                "excel",
+            ],
+        )
+
+        # relations_equal
         check_relations_equal(
             project.adapter,
             [
                 "tt1",
-                "source_model1",
+                "sqlalchemy1",
             ],
         )
 
+        # relations_equal
         check_relations_equal(
             project.adapter,
             [
                 "tt2",
-                "source_model2",
+                "sqlalchemy2",
             ],
         )
